app-container: try out "root-container" naming
Based on feeback from a number of folks the overloading of the term
"container" has caused quite a bit of confusion. This PR tries the term
"root-container" on for size. Which if you think of cgroups as a
heirarchy like the filesystem (which they are) makes a lot of sense.

This is an alternative to using a new term such as "pod", "group" or
"cell". The idea is that we don't call `/` on the UNIX filesystem
anything special, we call it the "root-directory". So, lets see if we
can try this on for size here. There is a root-container and one or more
sub-containers. Apps execute as a direct sub-container of the
root-container.

https://github.com/coreos/rocket/issues/151